fixed network issue
prevents the need of using '--net host' when launching the container. The python monitor will now get the IP of the docker interface and use this to communicate back to the native host server.

It is important to alter the firewall of the host to allow incoming connections from the docker0 interface using port 80